pyproject.toml
[project]
name = "esthetician-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "python-telegram-bot>=22.3",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]

[tool.ruff]
include = ["./*.py", "./**/*.py"]
line-length = 100
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN401", # (any-type)
    "COM812", # (missing-trailing-comma)
    "D100",  # (undocumented-public-module)
    "D101", # (undocument-public-class)
    "D102", # (undocumented-public-method)
    "D103", # (undocumented-public-function)
    "D104", # (undocumented-public-package)
    "D105",  # (undocumented-magic-method)
    "D107",  # (undocumented-public-init)
    "D203", # (one-blank-line-before-class)
    "D213", # (multi-line-summary-second-line)
]
extend-select = ["I"]